requirements.txt
discord.py
aiohttp
aiosqlite
python-dotenv
Pillow
Flask